Fixed "incorrect signature" issue for user placeholder image. Performed code cleanup.
