+++ app/src/main/java/com/sivan/sbaceman/presentation/home_screen/components/SpaceTimings.kt
@@ -17,6 +17,6 @@ fun SpaceTimings(modifier: Modifier, time: String) {
                 color = MaterialTheme.colorScheme.onSecondaryContainer
             ),
 
-            )
+        )
     }
 }
